+++ components/__init__.py
@@ -1,0 +1,3 @@
+from .header import get_header, get_logo, get_menu
+from .table import make_dash_table
+from .printButton import print_button
