+++ components/printButton.py
@@ -1,0 +1,5 @@
+import dash_html_components as html
+
+def print_button():
+    printButton = html.A(['Print PDF'],className="button no-print print",style={'position': "absolute", 'top': '-40', 'right': '0'})
+    return printButton
