+++ components/table.py
@@ -1,0 +1,11 @@
+import dash_html_components as html
+
+def make_dash_table(df):
+    ''' Return a dash definitio of an HTML table for a Pandas dataframe '''
+    table = []
+    for index, row in df.iterrows():
+        html_row = []
+        for i in range(len(row)):
+            html_row.append(html.Td([row[i]]))
+        table.append(html.Tr(html_row))
+    return table
